feat: Generalize agent pipeline and add specialized technique generators
Summary:
========

1. Specialized Technique Generators
   - Added `run_obfuscation_generator.py`: Targets anti-analysis, signature evasion, and pattern breaking techniques.
   - Added `run_badbyte_generator.py`: Specifically targets novel restricted byte elimination strategies.
   - Added `run_profile_generator.py`: Dynamically extracts restricted byte sets from `src/badbyte_profiles.h` to propose profile-specific strategies (e.g., alphanumeric-only, url-safe).

2. Agent Pipeline Generalization
   - `agents/strategy_discovery_agent.py`: Updated to be mode-aware (bad_byte vs. obfuscation) with context-sensitive summary prompts.
   - `agents/technique_proposal_agent.py`: Enhanced to support 'obfuscation' and 'profile' modes. Proposals now respect specific bad-byte constraints for profile-specific targets.
   - `agents/code_generation_agent.py`: Updated to handle mode-specific hard constraints and role-focus in C code generation.

3. Compiler Warning Fixes
   - `src/vex_escape_byte_remapping_for_simd_mov_strategies.c`: Fixed `-Wunused-parameter` for `orig_bytes` in `can_remap_with_nop`.
   - `src/vex_escape_byte_remapping_strategies.c`: Fixed `-Wsign-compare` in `generate_vex_escape_remapping` by casting `vex_type` to `size_t`.

4. Documentation Updates
   - `README.md`: Updated "Agent Menagerie" section with specialized generator usage.
   - `docs/USAGE.md`: Expanded "What's New in v4.3" with details on specialized obfuscation and profile-specific generators.

Validated: `make` clean build successful with zero warnings in modified files. Profile-specific byte extraction verified with `run_profile_generator.py` logic test.

+++ src/vex_escape_byte_remapping_strategies.c
@@ -265,10 +265,10 @@ static void generate_vex_escape_remapping(struct buffer *b, cs_insn *insn) {
     }
     
     /* Add the actual opcode byte(s) after VEX */
-    if (orig_size > vex_type) {
+    if (orig_size > (size_t)vex_type) {
         uint32_t opcode_part = 0;
-        for (size_t i = vex_type; i < orig_size && i < vex_type + 4; i++) {
-            opcode_part |= (orig_bytes[i] << ((i - vex_type) * 8));
+        for (size_t i = (size_t)vex_type; i < orig_size && i < (size_t)vex_type + 4; i++) {
+            opcode_part |= ((uint32_t)orig_bytes[i] << ((i - (size_t)vex_type) * 8));
         }
         packed_vex |= opcode_part;
     }
